+++ tests/boost_fatal_error.cpp
@@ -1,0 +1,10 @@
+#include <stdexcept>
+#define BOOST_TEST_MODULE MyTest
+#include <boost/test/included/unit_test.hpp>
+
+
+BOOST_AUTO_TEST_CASE( test_failure_1 )
+{
+    BOOST_REQUIRE( 2 * 3 == 5 );
+}
+