+++ tests/test_pytest_cpp.py
@@ -119,6 +119,17 @@ def test_boost_failure(exes):
     assert fail2.get_file_reference() == ("boost_failure.cpp", 14)
 
 
+def test_boost_fatal_error(exes):
+    facade = BoostTestFacade()
+    failures = facade.run_test(exes.get('boost_fatal_error'), '<unused>')
+    assert len(failures) == 1
+
+    fail1, = failures
+    colors = ('red', 'bold')
+    assert fail1.get_lines() == [('critical check 2 * 3 == 5 failed', colors)]
+    assert fail1.get_file_reference() == ("boost_fatal_error.cpp", 8)
+
+
 def test_boost_error(exes):
     facade = BoostTestFacade()
     failures = facade.run_test(exes.get('boost_error'), '<unused>')